Added automatic CSV parsing and cell creations
Using Mark Price’s CSV parser, the Pokemon data was collected from a
CSV file in the form of dictionaries.

Then a ‘poke’ constant was created that as assigned to the Pokemon at a
specific row of the CSV dictionary.

The newly created Pokemon object was added to the array of Pokemon.